Created new project with tests and fixed tests' names
