+++ cs/testing.sln
@@ -9,7 +9,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HomeExercises", "HomeExerci
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Challenge", "Challenge\Challenge.csproj", "{BB8C2EFB-6AE9-45BD-8468-829D5AB79DCB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NumberValidatorTests", "HomeExercises\NumberValidatorTests\NumberValidatorTests.csproj", "{EAAB1A23-ECBA-4962-9DE1-72BE6AD480E1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HomeWorkTests", "HomeWorkTests\HomeWorkTests.csproj", "{1813E583-E2AE-48B9-8639-1CD46BA9E511}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,10 +29,10 @@ Global
 		{BB8C2EFB-6AE9-45BD-8468-829D5AB79DCB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB8C2EFB-6AE9-45BD-8468-829D5AB79DCB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BB8C2EFB-6AE9-45BD-8468-829D5AB79DCB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EAAB1A23-ECBA-4962-9DE1-72BE6AD480E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EAAB1A23-ECBA-4962-9DE1-72BE6AD480E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EAAB1A23-ECBA-4962-9DE1-72BE6AD480E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EAAB1A23-ECBA-4962-9DE1-72BE6AD480E1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1813E583-E2AE-48B9-8639-1CD46BA9E511}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1813E583-E2AE-48B9-8639-1CD46BA9E511}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1813E583-E2AE-48B9-8639-1CD46BA9E511}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1813E583-E2AE-48B9-8639-1CD46BA9E511}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
